moving globals into shinstance...
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@1198 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kash/alias.c
@@ -104,7 +104,7 @@ setalias(shinstance *psh, char *name, char *val)
 	ap->val = savestr(val);
 #else /* hack */
 	{
-	int len = strlen(val);
+	size_t len = strlen(val);
 	ap->val = ckmalloc(len + 2);
 	memcpy(ap->val, val, len);
 	ap->val[len] = ' ';	/* fluff */
